Conflicts:
	src/main/webapp/WEB-INF/views/jsp/mainFrame.jsp
	src/main/webapp/WEB-INF/views/jsp/treeView.jsp

+++ src/test/java/maple_tiger_hr/service/AbstractHRServiceTests.java
@@ -145,9 +145,16 @@ public class AbstractHRServiceTests {
     
     @Test
     public void shouldDeleteDepartmentById(){
-                            
-        this.hrService.deleteDepartmentById(7);
+        
+        Department dept = this.hrService.findDepartmentById(7);
+        if(dept != null){
+            this.hrService.deleteDepartmentById(7);
+        }        
             
+        TreeNode<Department> tree = this.hrService.getTreeFromDepartments();
+        System.out.println("tree: After deletion --------------------------------------------");
+        System.out.println(tree);
+        System.out.println("tree: --------------------------------------------");
         //Department dept = this.hrService.findDepartmentById(7);
         //assertThat(dept).isNull();
